Add files via upload
A Bayesian Network (BN) is a probabilistic graphical model that represents a set of variables and their conditional dependencies via a directed acyclic graph (DAG). When considering medical data, such as data related to heart disease, Bayesian networks can be an effective tool for modeling the uncertainty and relationships between various risk factors, symptoms, and possible outcomes. Here are some uses and considerations for implementing a Bayesian network in the context of heart disease. This cannot handle the heavy data